+++ packages/plugins/Legend/src/index.ts
@@ -1,10 +1,10 @@
 import Vue from 'vue'
-import { PluginOptions } from '@polar/lib-custom-types'
+import { LegendConfiguration } from '@polar/lib-custom-types'
 
 import { Legend } from './components'
 import locales from './locales'
 
-export default (options: PluginOptions) => (instance: Vue) =>
+export default (options: LegendConfiguration) => (instance: Vue) =>
   instance.$store.dispatch('addComponent', {
     name: 'legend',
     plugin: Legend,